pyproject.toml
[project]
name = "reface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.4.0",
    "black>=25.1.0",
    "cmake>=3.31.4",
    "diffusers>=0.32.2",
    "dlib>=19.24.6",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "isort>=6.0.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pre-commit>=4.1.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.7",
]
